API

CommonTaskExercise.denton_interpolation — Method

denton_interpolation(highfreq::AbstractMatrix{<:Real},
                     lowfreq::AbstractVector{<:Real})::Matrix{Float64}

Returns the interpolation using the Proportional Denton Method. The implementation is taken from the Quarterly National Accounts Manual – 2017 Edition (Chapter 6, pages 121-122).

It assumes no gaps in the series.

CommonTaskExercise.read_xlsx_sheet — Method

read_xlsx_sheet(file::XLSXFile, sheetname::AbstractString)::DataFrame

Returns the data from the sheet for each series from 2010 forward.

Assumes:

A1:A7 contains the metadata header

Column names are given by the 8th row

A: Line

B: Description

C: PublishCd

D and on: Observation Period (e.g., 2010, 2010Q1, 2010M01)

Data starts on the 9th row

Non-numeric values (e.g., ".....") are interpreted as 0.